+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Craft Commerce
 
+## Unreleased
+
+- Fixed a bug where carts created as a guest weren’t being associated with the user when they logged in. ([#3445](https://github.com/craftcms/commerce/issues/3445))
+
 ## 4.5.3 - 2024-04-02
 
 - The “Postcode Condition Formula” condition rule now allows multi-line input. ([#3147](https://github.com/craftcms/commerce/issues/3147)